Fix in User module

+++ modules/Market/Database/Migrations/2022_02_16_195229_create_guaranties_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->id();
             $table->string('title')->unique();
             $table->unsignedBigInteger('product_id');
-            $table->decimal('price_increase', 10, 1)->default(0);
+            $table->unsignedInteger('price_increase')->default(0);
             $table->tinyInteger('status')->default(0);
             $table->timestamps();
             $table->softDeletes();
