+++ modules/Payment/Database/Migrations/2022_02_17_155015_create_payments_table.php
@@ -15,7 +15,7 @@ class CreatePaymentsTable extends Migration
     {
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
-            $table->decimal('amount', 20, 3);
+            $table->unsignedInteger('amount');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('vendor_id');
             $table->tinyInteger('status')->default(0);
